+++ src/consumer/bit-map/component-map.js
@@ -201,7 +201,7 @@ export default class ComponentMap {
         componentPaths: [trackDirRelative || '.'],
         id: id.toString(),
         override: false, // this makes sure to not override existing files of componentMap
-        writeToBitMap: false,
+        trackDirFeature: true,
         origin: this.origin
       };
       const numOfFilesBefore = this.files.length;